requirements.txt
sec_cik_mapper
streamlit
BeautifulSoup
pillow